feat: refactor and consume fsRoutes as plugins
