+++ examples/start.ts
@@ -1,6 +1,6 @@
-import { LimetteApp } from "@limette/core";
+import { App } from "@limette/core";
 
-export const app = new LimetteApp();
+export const app = new App();
 
 if (import.meta.main) {
   app.listen({ port: 8000 });
