+++ src/mod.ts
@@ -1,6 +1,7 @@
-// export { LimetteApp } from "./server/app.ts";
-export { LimetteApp } from "./server/app.ts";
+export { App } from "./server/app.ts";
 export { dev } from "./dev/dev.ts";
+export { fsRoutes } from "./server/fs-routes.ts";
+export { staticFiles } from "./server/static-files.ts";
 export type { AppRootOptions } from "./server/ssr.ts";
 export type { Handlers } from "./server/router.ts";
 export type { Context } from "./server/app.ts";
