+++ src/server/router.ts
@@ -1,9 +1,6 @@
 import type { MiddlewareFn } from "./middleware.ts";
-import { getRoutes, getAppTemplate } from "../dev/build.ts";
 import "./ssr.ts";
 import type { Context } from "./app.ts";
-import type { LimetteApp } from "./app.ts";
-import { renderContent } from "./ssr.ts";
 
 export type Method = "HEAD" | "GET" | "POST" | "PATCH" | "PUT" | "DELETE";
 
@@ -21,13 +18,6 @@ export interface Route {
   handlers: MiddlewareFn[];
 }
 
-export interface GetRouterOptions {
-  buildAssets?: boolean;
-  devMode?: boolean;
-  staticRoutes?: boolean;
-  loadFs?: (path: string) => Promise<unknown>;
-}
-
 export type Handlers = {
   GET?(ctx: Context): Response | Promise<Response>;
   POST?(ctx: Context): Response | Promise<Response>;
@@ -36,11 +26,6 @@ export type Handlers = {
   PATCH?(ctx: Context): Response | Promise<Response>;
   OPTIONS?(ctx: Context): Response | Promise<Response>;
   HEAD?(ctx: Context): Response | Promise<Response>;
-};
-
-export type ComponentContext = {
-  params: Context["params"];
-  data: unknown;
 };
 
 export class UrlPatternRouter {
@@ -104,97 +89,4 @@ export class UrlPatternRouter {
 
     return result;
   }
-}
-
-export async function setRoutes(app: LimetteApp, options: GetRouterOptions) {
-  const [routes, AppRoot] = await Promise.all([
-    getRoutes(options),
-    getAppTemplate(options),
-  ]);
-
-  if (!AppRoot) {
-    throw new Error(
-      "You need to create an AppRoot (_app.ts/js) to render a page."
-    );
-  }
-
-  // Serve static files from memory on dev mode
-  if (options?.devMode) {
-    app.get("/_limette/js/chunk-:id.js", (ctx) => {
-      const { id } = ctx.params;
-      const route = routes.find((r) => r.id === id);
-
-      return new Response(route?.jsAssetContent?.contents, {
-        headers: {
-          "Content-Type": "application/javascript; charset=UTF-8",
-        },
-      });
-    });
-
-    app.get("/_limette/css/tailwind-:id.css", (ctx) => {
-      const { id } = ctx.params;
-      const route = routes.find((r) => r.id === id);
-
-      return new Response(route?.cssAssetContent, {
-        headers: {
-          "Content-Type": "text/css; charset=UTF-8",
-        },
-      });
-    });
-  }
-
-  routes.map((route) => {
-    const middlewares = route.middlewares
-      .map((module) => module?.handler)
-      .flat();
-
-    // Register custom handlers
-    if (route.routeModule?.handler) {
-      for (const [method, handlerFn] of Object.entries(
-        route.routeModule.handler
-      )) {
-        const routeHandler = async (ctx: Context) => {
-          ctx.render = async (data: ComponentContext["data"]) => {
-            if (!route.routeModule?.default) {
-              throw new Error(
-                "No component was provided. Make sure you export a component as default to be redered."
-              );
-            }
-
-            const content = await renderContent(AppRoot, route, ctx, data);
-
-            return new Response(content, {
-              status: 200,
-              statusText: "OK",
-              headers: new Headers({ "Content-Type": "text/html" }),
-            });
-          };
-
-          return await handlerFn(ctx);
-        };
-
-        // Register route
-        app[method.toLocaleLowerCase() as Lowercase<Method>](
-          route.path,
-          ...middlewares,
-          routeHandler
-        );
-      }
-    }
-
-    // Default behaviour if no GET handler is provided
-    if (route.routeModule?.default && !route.routeModule?.handler?.GET) {
-      const routeHandler = async (ctx: Context) => {
-        const content = await renderContent(AppRoot, route, ctx);
-
-        return new Response(content, {
-          headers: {
-            "Content-Type": "text/html",
-          },
-        });
-      };
-
-      app.get(route.path, ...middlewares, routeHandler);
-    }
-  });
 }
